Move internal/database/oauthtoken to internal/oauthtoken
