+++ internal/authz/providers/azuredevops/BUILD.bazel
@@ -10,11 +10,11 @@ go_library(
         "//internal/authz",
         "//internal/authz/types",
         "//internal/database",
-        "//internal/database/oauthtoken",
         "//internal/extsvc",
         "//internal/extsvc/auth",
         "//internal/extsvc/azuredevops",
         "//internal/licensing",
+        "//internal/oauthtoken",
         "//internal/types",
         "//lib/errors",
         "@com_github_sourcegraph_log//:log",
