+++ internal/authz/providers/bitbucketcloud/BUILD.bazel
@@ -13,12 +13,12 @@ go_library(
         "//internal/authz",
         "//internal/authz/types",
         "//internal/database",
-        "//internal/database/oauthtoken",
         "//internal/extsvc",
         "//internal/extsvc/auth",
         "//internal/extsvc/bitbucketcloud",
         "//internal/httpcli",
         "//internal/licensing",
+        "//internal/oauthtoken",
         "//internal/types",
         "//lib/errors",
         "//schema",
