+++ internal/authz/providers/github/BUILD.bazel
@@ -15,13 +15,13 @@ go_library(
         "//internal/authz",
         "//internal/authz/types",
         "//internal/database",
-        "//internal/database/oauthtoken",
         "//internal/encryption/keyring",
         "//internal/extsvc",
         "//internal/extsvc/auth",
         "//internal/extsvc/github",
         "//internal/extsvc/github/auth",
         "//internal/licensing",
+        "//internal/oauthtoken",
         "//internal/rcache",
         "//internal/types",
         "//lib/errors",
