+++ internal/authz/providers/gitlab/BUILD.bazel
@@ -15,13 +15,13 @@ go_library(
         "//internal/authz",
         "//internal/authz/types",
         "//internal/database",
-        "//internal/database/oauthtoken",
         "//internal/extsvc",
         "//internal/extsvc/auth",
         "//internal/extsvc/gitlab",
         "//internal/featureflag",
         "//internal/httpcli",
         "//internal/licensing",
+        "//internal/oauthtoken",
         "//internal/types",
         "//lib/errors",
         "//schema",
