+++ internal/oauthtoken/BUILD.bazel
@@ -4,7 +4,7 @@ load("//dev:go_defs.bzl", "go_test")
 go_library(
     name = "oauthtoken",
     srcs = ["oauth_token_helper.go"],
-    importpath = "github.com/sourcegraph/sourcegraph/internal/database/oauthtoken",
+    importpath = "github.com/sourcegraph/sourcegraph/internal/oauthtoken",
     visibility = ["//:__subpackages__"],
     deps = [
         "//internal/database",
